+++ src/main/kotlin/com/barostartbe/domain/mentee/controller/MenteeController.kt
@@ -1,7 +1,10 @@
 package com.barostartbe.domain.mentee.controller
 
 import com.barostartbe.domain.mentee.dto.GetMenteeBasicInfoResponseDto
+import com.barostartbe.domain.mentee.dto.GetMenteeDashboardResponseDto
+import com.barostartbe.domain.mentee.dto.GetMentoMainDashboardResponseDto
 import com.barostartbe.domain.mentee.usecase.MenteeQueryUseCase
+import com.barostartbe.domain.mentor.entity.Mentor
 import com.barostartbe.domain.user.entity.User
 import com.barostartbe.global.response.ApiResponse
 import com.barostartbe.global.response.type.SuccessCode
@@ -18,4 +21,17 @@ class MenteeController(
     override fun getMenteeInfo(@PathVariable menteeId: Long, @AuthenticationPrincipal mentor: User): ResponseEntity<ApiResponse<GetMenteeBasicInfoResponseDto>> {
         return ApiResponse.success(SuccessCode.REQUEST_OK, menteeQueryUseCase.getMenteeInfo(mentor.id!!, menteeId))
     }
+
+    override fun getMenteeDetailsDashboard(
+        menteeId: Long,
+        searchType: String,
+        date: String?,
+        mentor: User
+    ): ResponseEntity<ApiResponse<GetMenteeDashboardResponseDto>> {
+        val response = menteeQueryUseCase.getMenteeDashboard(mentor as Mentor, menteeId, searchType, date)
+        return ApiResponse.success(SuccessCode.REQUEST_OK, response)
+    }
+
+    override fun getMentorMainDashboard(mentorId: Long): ResponseEntity<ApiResponse<GetMentoMainDashboardResponseDto>>
+        = ApiResponse.success(SuccessCode.REQUEST_OK, menteeQueryUseCase.getMenteesTodayDetails(mentorId))
 }
